Validación de mantenimientos de seccion

+++ includes/eliminar_seccion.php
@@ -6,20 +6,24 @@
         $se_id=$_POST['se_id'];        
         if($se_id=="")
         {
-            echo "No hay seccion ingresada";
+            echo "ERROR: Llene los campos obligatorios!";
         }
         else
         {
-            $peticion= "delete from seccion where id =('".$se_id."')";
-            $resultado=mysqli_query($conexion,$peticion);
-            if($resultado)
-            {
-                echo "eliminada";
-            }
-            else
-            {
-                echo "error";
-            }       
+            if (is_numeric($se_id)) {
+                $peticion= "delete from seccion where id =('".$se_id."')";
+                $resultado=mysqli_query($conexion,$peticion);
+                if($resultado)
+                {
+                    echo "ADVERTENCIA: La sección se ha eliminado correctamente!";
+                }
+                else
+                {
+                    echo "Error: Error al intentar eliminar la sección!";
+                }  
+            }else{
+                echo "ERROR: Conflicto al eliminar el ID de la sección!";
+            }      
         }
          
     }
